feat(admin): 管理ツールをダッシュボードに集約 (埋め込み、別画面へ飛ばさない)
Closes #428
Epic #416

「バラバラの画面に飛ばしたら意味ない」(オーナー指摘)。#417 のダッシュボードは既存ツールへ
リンクしていたが、埋め込みで集約する。

- ServerOAuthAdmin / WorldResetAdmin を settings.tsx から components/admin/ へ抽出 (export)。
- パワー付与を spirit.tsx から components/admin/power-grant-admin.tsx へ抽出 (points を自前ロード
  する自己完結コンポーネントに)。
- /admin ダッシュボードに「ツール」として PowerGrantAdmin / WorldResetAdmin / DebugBattleSim /
  ServerOAuthAdmin を埋め込み表示。
- settings.tsx から両コンポーネント + render + 未使用 import を撤去 (「ダッシュボードを開く」
  導線は残す)。spirit.tsx からパワー付与 + 模擬戦 render + 未使用 import/state を撤去。

web typecheck / test 346 / build 緑。
