+++ apps/web/src/routes/spirit.tsx
@@ -21,8 +21,6 @@ import { useRuntimeConfig } from '@/components/config-provider';
 import { applyPromptTemplate } from '@/lib/prompt-template';
 import { bumpPower, loadPointsState, SUMMON_THRESHOLD, type PointsState } from '@/lib/points';
 import { WORLD_PREVIEW_ENABLED } from '@/lib/world-preview';
-import { DebugBattleSim } from '@/components/debug-battle-sim';
-import { isAdminDid } from '@/lib/runtime-config';
 
 /**
  * 精霊ブルスコンのページ。
@@ -52,8 +50,6 @@ export function Spirit() {
   const [questXp, setQuestXp] = useState(0);
   const [loaded, setLoaded] = useState(false);
   const [ritualOpen, setRitualOpen] = useState(false);
-  // 管理者用パワー付与の書込み中フラグ (dev 限定・多重押し防止)
-  const [grantingPower, setGrantingPower] = useState(false);
 
   const agent = session.agent ?? null;
   const did = session.did ?? null;
@@ -278,54 +274,8 @@ export function Spirit() {
             </section>
           )}
 
-          {/* 管理者用: dev でパワーを付与してテストする。VITE_ADMIN_DIDS の DID かつ
-              プレビュー環境 (dev/local) のときだけ表示。本番遮断は WORLD_PREVIEW_ENABLED
-              (本番ビルドでは false → dead-code) が担い、isAdminDid は dev 内での権限分離。
-              加算は残高に純増する salePowerEarned に乗せる (viaPosts は召喚進捗ゲージ・
-              toSummon も動かす別セマンティクスなので、残高だけ盛る用途では使わない)。 */}
-          {WORLD_PREVIEW_ENABLED && agent && did && isAdminDid(did) && (
-            <section style={{ marginTop: '1em', textAlign: 'center' }}>
-              <div
-                style={{
-                  display: 'inline-flex',
-                  gap: '0.5em',
-                  alignItems: 'center',
-                  fontSize: '0.78em',
-                  border: '1px dashed var(--color-muted)',
-                  borderRadius: 6,
-                  padding: '0.45em 0.7em',
-                  color: 'var(--color-muted)',
-                }}
-              >
-                <span>{grantingPower ? '管理者 · 付与中…' : `管理者 · パワー ${points.balance}`}</span>
-                {[100, 1000].map((n) => (
-                  <button
-                    key={n}
-                    type="button"
-                    disabled={grantingPower}
-                    onClick={async () => {
-                      setGrantingPower(true);
-                      try {
-                        await bumpPower(agent, did, { salePowerEarned: n });
-                        const next = await loadPointsState(agent, did);
-                        setPoints(next);
-                      } catch (e) {
-                        console.warn('admin power grant failed', e);
-                      } finally {
-                        setGrantingPower(false);
-                      }
-                    }}
-                    style={{ fontSize: '1em', padding: '0.2em 0.6em' }}
-                  >
-                    +{n}
-                  </button>
-                ))}
-              </div>
-            </section>
-          )}
-
-          {/* 管理者デバッグ模擬戦 (issue #414)。dev + 管理者のみ。 */}
-          {WORLD_PREVIEW_ENABLED && did && isAdminDid(did) && <DebugBattleSim />}
+          {/* 管理者ツール (パワー付与・模擬戦・リセット・連携) は管理ダッシュボード /admin に
+              集約した (以前はここ。バラバラに置かない — オーナー指摘 2026-07-20)。 */}
         </>
       )}
